Initial V1 overhaul
The previous version was old and bloated, with an unnecessary Create
React App base, CSS Modules, and overall outdated package versions.

This commit removes many of those base folders, files, and scripts and
sets a more clean foundation to builder a newer modern version.

Note that existing components were left in as-is, so the library doesn't
build because they're expecting Classnames and CSS Modules.